test: merge bedrock service pagination tests into main test file
